requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
jupyter>=1.0.0
pandas>=1.3.0
scipy>=1.7.0